refactor: 수신자(Receiver) 기능 모듈 분리 및 관련 도메인 로직 이전
- `feature:receiver` 하위에 data, domain, presentation 모듈을 신설하고 `settings.gradle.kts`에 등록
- `feature:afternote:domain`에 포함되어 있던 수신자 인증(`ReceiverAuth`), 본인 확인(`IdentityVerification`), 증빙 서류 업로드 관련 도메인 모델 및 리포지토리 인터페이스를 `feature:receiver:domain`으로 이동
- 패키지 경로 변경(`com.afternote.feature.afternote.domain.model.receiver` -> `com.afternote.feature.receiver.domain.model` 등)에 따른 프로젝트 전반의 임포트 경로 수정
- `feature:afternote` 모듈(data, presentation)에서 분리된 `feature:receiver:domain` 모듈을 참조하도록 의존성 설정 업데이트
- 신규 모듈(`feature:receiver:data`, `feature:receiver:domain`)의 `build.gradle.kts` 및 기초 매니페스트 구성 추가
