+++ feature/afternote/data/src/main/kotlin/com/afternote/feature/afternote/data/di/AfternoteReceiverRepositoryModule.kt
@@ -2,8 +2,8 @@ package com.afternote.feature.afternote.data.di
 
 import com.afternote.feature.afternote.data.repositoryimpl.receiver.IdentityVerificationRepositoryImpl
 import com.afternote.feature.afternote.data.repositoryimpl.receiver.ReceiverRepositoryImpl
-import com.afternote.feature.afternote.domain.repository.receiver.IdentityVerificationRepository
 import com.afternote.feature.afternote.domain.repository.receiver.ReceiverRepository
+import com.afternote.feature.receiver.domain.repository.IdentityVerificationRepository
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
