+++ feature/afternote/data/src/main/kotlin/com/afternote/feature/afternote/data/repositoryimpl/receiver/IdentityVerificationRepositoryImpl.kt
@@ -6,7 +6,7 @@ import androidx.datastore.preferences.core.booleanPreferencesKey
 import androidx.datastore.preferences.core.edit
 import androidx.datastore.preferences.core.emptyPreferences
 import com.afternote.feature.afternote.data.di.IdentityVerificationDataStore
-import com.afternote.feature.afternote.domain.repository.receiver.IdentityVerificationRepository
+import com.afternote.feature.receiver.domain.repository.IdentityVerificationRepository
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.catch
 import kotlinx.coroutines.flow.map
