+++ feature/afternote/presentation/src/main/kotlin/com/afternote/feature/afternote/presentation/receiver/navigation/model/ReceiverRoute.kt
@@ -66,7 +66,7 @@ sealed interface ReceiverRoute {
     /**
      * 본인 확인 이메일 인증 화면(designs 3·4). 인증 시작하기 → 진입.
      *
-     * 인증 성공 시 [com.afternote.feature.afternote.domain.repository.receiver.IdentityVerificationRepository]
+     * 인증 성공 시 [com.afternote.feature.receiver.domain.repository.IdentityVerificationRepository]
      * 캐시가 켜져 이후 동일 사용자(폰)에서는 마스터 키로 직진. DataStore 영구 저장이라 앱 재시작 후에도 유지.
      */
     @Serializable
