+++ feature/receiver/data/build.gradle.kts
@@ -3,15 +3,13 @@ plugins {
 }
 
 android {
-    namespace = "com.afternote.feature.afternote.data"
+    namespace = "com.afternote.feature.receiver.data"
 }
 
 dependencies {
     implementation(libs.androidx.datastore.preferences)
-    implementation(libs.androidx.paging.runtime)
     implementation(projects.core.common)
     implementation(projects.core.domain)
     implementation(projects.core.model)
-    implementation(projects.feature.afternote.domain)
     implementation(projects.feature.receiver.domain)
 }
